// client/src/app/models/Course.ts
export interface Course {
  id?: number;
  code?: number;
  name?: string;
  level?: number;
  day_1?: string;
  day_2?: string;
  time_1?: string;
  time_2?: string;
  generation?: string;
  created_at?: Date;
  modified_on?: Date;
  instructor_id?: number;
  status?: string;
  count?: number;
  instructor_name?: string;
  instructorName?: string;
  instructor_last_name?: string;
  instructor_email?: string;
  start?: Date;
  stake?: string;
}